Updated setting scroll for dynamic loading
If the website dynamically loads when you scroll down, the program would
stop at the loading place: thus not going where it should.
Now, if the new scroll location is not the desired one, it waits
a second, then scrolls again.
This repeats until the desired scroll location is reached!